Refactoring
The `Meaning` typeclass for doing interpretations is not longer
used. Lean's dot-notation is infinitely more convenient and
faster.

Also now if a proposition is `True` it basic logical operations
like `And` now get short-circuited without performing any actual
computation in lean. The idea is that lemmas that are compiled to
lean are more or less readable.